+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [?](https://github.com/sameday-courier/php-sdk/compare/1.1.0...?)
+
+### Added
+
+- Download PDF for an existing AWB
+
 ## [1.1.0](https://github.com/sameday-courier/php-sdk/compare/1.0.0...1.1.0) - 2019-01-15
 
 ### Fixed